requirements.txt
numpy
opencv-python
pyserial
pylibfreenect2
